Multi-Agent RAG Scientific Paper Verifier/requirements.txt
sentence-transformers
faiss-cpu
pandas
numpy
scikit-learn
streamlit